% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_empty_dots}
\alias{assert_empty_dots}
\title{Assert that dots are empty}
\usage{
assert_empty_dots(...)
}
\arguments{
\item{...}{(\code{any})\cr
Ellipsis arguments to check.}
}
\value{
\code{NULL} invisibly if empty, otherwise throws an error.
}
\description{
Assert that dots are empty
}
