% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_base64}
\alias{is_base64}
\title{Check if an argument is a base64 string}
\usage{
is_base64(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid non-empty base64 string, \code{FALSE} otherwise.
}
\description{
Check if an argument is a base64 string
}
\examples{
is_base64("SGVsbG8gV29ybGQ=")
}
