% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_bic}
\alias{is_bic}
\title{Check if an argument is a BIC/SWIFT code string}
\usage{
is_bic(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid BIC/SWIFT code string, `FALSE` otherwise.
}
\description{
Check if an argument is a BIC/SWIFT code string
}
\examples{
is_bic("DEUTDEFF")
}
\references{
<https://en.wikipedia.org/wiki/ISO_9362>
<https://knowledge.xmldation.com/support/iso20022/general_rules/bic>
}
