% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_doi}
\alias{is_doi}
\title{Check if an argument is a DOI string}
\usage{
is_doi(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid DOI string, `FALSE` otherwise.
}
\description{
Check if an argument is a DOI string
}
\examples{
is_doi("10.1038/nphys1170")
}
\references{
<https://www.doi.org/the-identifier/what-is-a-doi/>
}
