% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_email}
\alias{is_email}
\title{Check if an argument is an email address string}
\usage{
is_email(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid email address string, \code{FALSE} otherwise.
}
\description{
Check if an argument is an email address string
}
\examples{
is_email("user@example.com")
}
