% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_hex}
\alias{is_hex}
\title{Check if an argument is a hexadecimal string}
\usage{
is_hex(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid non-empty hexadecimal string, `FALSE`
  otherwise.
}
\description{
Check if an argument is a hexadecimal string
}
\examples{
is_hex("deadbeef")
}
