% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_hostname}
\alias{is_hostname}
\title{Check if an argument is a hostname string}
\usage{
is_hostname(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid hostname string, `FALSE` otherwise.
}
\description{
Check if an argument is a hostname string
}
\examples{
is_hostname("example.com")
}
\references{
<https://en.wikipedia.org/wiki/Hostname>
}
