% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_iban}
\alias{is_iban}
\title{Check if an argument is a valid IBAN string}
\usage{
is_iban(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid IBAN string, \code{FALSE} otherwise.
}
\description{
Validates IBAN format including MOD-97-10 check digit verification (ISO/IEC 7064).
}
\examples{
is_iban("GB29NWBK60161331926819")
}
\references{
\url{https://en.wikipedia.org/wiki/International_Bank_Account_Number}
}
