% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_ipv4}
\alias{is_ipv4}
\title{Check if an argument is an IPv4 address string}
\usage{
is_ipv4(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid IPv4 address string, \code{FALSE} otherwise.
}
\description{
Check if an argument is an IPv4 address string
}
\examples{
is_ipv4("192.168.1.1")
}
