% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_issn}
\alias{is_issn}
\title{Check if an argument is an ISSN string}
\usage{
is_issn(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid ISSN string, `FALSE` otherwise.
}
\description{
Validates ISSN format including check digit verification.
}
\examples{
is_issn("0378-5955")
}
\references{
<https://www.loc.gov/issn/check.html>
}
