% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_lei}
\alias{is_lei}
\title{Check if an argument is a valid LEI string}
\usage{
is_lei(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid LEI string, \code{FALSE} otherwise.
}
\description{
Validates LEI (Legal Entity Identifier) format including MOD-97-10 check digit verification
(ISO/IEC 7064).
}
\references{
\url{https://en.wikipedia.org/wiki/Legal_Entity_Identifier}
\url{https://www.govinfo.gov/content/pkg/CFR-2016-title12-vol8/xml/CFR-2016-title12-vol8-part1003-appC.xml}
}
