% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_mac}
\alias{is_mac}
\title{Check if an argument is a MAC address string}
\usage{
is_mac(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid MAC address string, `FALSE` otherwise.
}
\description{
Check if an argument is a MAC address string
}
\examples{
is_mac("00:1B:44:11:3A:B7")
}
