% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_nanoid}
\alias{is_nanoid}
\title{Check if an argument is a Nano ID string}
\usage{
is_nanoid(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid Nano ID string, `FALSE` otherwise.
}
\description{
Check if an argument is a Nano ID string
}
\examples{
is_nanoid("V1StGXR8_Z5jdHi6B-myT")
}
