% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_orcid}
\alias{is_orcid}
\title{Check if an argument is an ORCID string}
\usage{
is_orcid(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid ORCID string, \code{FALSE} otherwise.
}
\description{
Check if an argument is an ORCID string
}
