% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_sha256}
\alias{is_sha256}
\title{Check if an argument is a SHA-256 hash string}
\usage{
is_sha256(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid SHA-256 hash string, \code{FALSE} otherwise.
}
\description{
Check if an argument is a SHA-256 hash string
}
