% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_url}
\alias{is_url}
\title{Check if an argument is a URL string}
\usage{
is_url(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid URL string, \code{FALSE} otherwise.
}
\description{
Check if an argument is a URL string
}
\examples{
is_url("https://example.com")
}
