% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_base64}
\alias{is_base64}
\title{Check if an argument is a base64 string}
\usage{
is_base64(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid base64 string, `FALSE` otherwise.
}
\description{
Check if an argument is a base64 string
}
\examples{
is_base64("SGVsbG8gV29ybGQ=")
}
