% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_base64url}
\alias{is_base64url}
\title{Check if an argument is base64url string}
\usage{
is_base64url(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid base64url string, `FALSE` otherwise.
}
\description{
Check if an argument is base64url string
}
\examples{
is_base64url("SGVsbG8gV29ybGQ")
}
