% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_bic}
\alias{is_bic}
\title{Check if an argument is a BIC/SWIFT code string}
\usage{
is_bic(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid BIC/SWIFT code string, \code{FALSE} otherwise.
}
\description{
Check if an argument is a BIC/SWIFT code string
}
\references{
\url{https://en.wikipedia.org/wiki/ISO_9362}
\url{https://knowledge.xmldation.com/support/iso20022/general_rules/bic}
}
