% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_cuid2}
\alias{is_cuid2}
\title{Check if an argument is a CUID2 string}
\usage{
is_cuid2(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid CUID2 string, \code{FALSE} otherwise.
}
\description{
Check if an argument is a CUID2 string
}
\examples{
is_cuid2("ckopqwooh000001la8mbi2im9")
}
