% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_cusip}
\alias{is_cusip}
\title{Check if an argument is a valid CUSIP string}
\usage{
is_cusip(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid CUSIP string, \code{FALSE} otherwise.
}
\description{
Validates CUSIP format including Luhn-variant check digit verification.
}
\examples{
is_cusip("037833100")
}
\references{
\url{https://en.wikipedia.org/wiki/CUSIP}
}
