% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_doi}
\alias{is_doi}
\title{Check if an argument is a DOI string}
\usage{
is_doi(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid DOI string, \code{FALSE} otherwise.
}
\description{
Check if an argument is a DOI string
}
\references{
\url{https://www.doi.org/the-identifier/what-is-a-doi/}
}
