% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_email}
\alias{is_email}
\title{Check if an argument is a email address string}
\usage{
is_email(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid email address string, `FALSE` otherwise.
}
\description{
Check if an argument is a email address string
}
\examples{
is_email("user@example.com")
}
