% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_email}
\alias{is_email}
\title{Check if an argument is a email address string}
\usage{
is_email(x)
}
\arguments{
\item{x}{(\code{any}) object to check.}
}
\description{
Check if an argument is a email address string
}
