% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_figi}
\alias{is_figi}
\title{Check if an argument is a valid FIGI string}
\usage{
is_figi(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid FIGI string, \code{FALSE} otherwise.
}
\description{
Validates FIGI (Financial Instrument Global Identifier) format including Luhn check digit
verification.
}
\examples{
is_figi("BBG000BLNNH6")
}
\references{
\url{https://en.wikipedia.org/wiki/Financial_Instrument_Global_Identifier}
\url{https://www.openfigi.com/assets/content/figi-check-digit-2173341b2d.pdf}
}
