% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_hostname}
\alias{is_hostname}
\title{Check if an argument is a hostname string}
\usage{
is_hostname(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid hostname string, \code{FALSE} otherwise.
}
\description{
Check if an argument is a hostname string
}
\references{
\url{https://en.wikipedia.org/wiki/Hostname}
}
