% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_isbn}
\alias{is_isbn}
\title{Check if an argument is an ISBN string}
\usage{
is_isbn(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid ISBN string, \code{FALSE} otherwise.
}
\description{
Validates ISBN-10 and ISBN-13 formats, including check digit verification.
Hyphens and spaces are allowed as separators.
}
\examples{
is_isbn("978-0-306-40615-7")
}
\references{
\url{https://en.wikipedia.org/wiki/International_Standard_Book_Number}
}
