% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_isin}
\alias{is_isin}
\title{Check if an argument is a valid ISIN string}
\usage{
is_isin(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid ISIN string, \code{FALSE} otherwise.
}
\description{
Validates ISIN (International Securities Identification Number) format including Luhn check
digit verification.
}
\references{
\url{https://en.wikipedia.org/wiki/International_Securities_Identification_Number}
}
