% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_issn}
\alias{is_issn}
\title{Check if an argument is an ISSN string}
\usage{
is_issn(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid ISSN string, \code{FALSE} otherwise.
}
\description{
Validates ISSN format including check digit verification.
}
\references{
\url{https://www.loc.gov/issn/check.html}
}
