% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_md5}
\alias{is_md5}
\title{Check if an argument is an MD5 hash string}
\usage{
is_md5(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid MD5 hash string, \code{FALSE} otherwise.
}
\description{
Check if an argument is an MD5 hash string
}
\examples{
is_md5("d41d8cd98f00b204e9800998ecf8427e")
}
