% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_nanoid}
\alias{is_nanoid}
\title{Check if an argument is a Nano ID string}
\usage{
is_nanoid(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid Nano ID string, \code{FALSE} otherwise.
}
\description{
Check if an argument is a Nano ID string
}
