% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_orcid}
\alias{is_orcid}
\title{Check if an argument is an ORCID string}
\usage{
is_orcid(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid ORCID string, \code{FALSE} otherwise.
}
\description{
Validates ORCID format including ISO/IEC 7064:2003, MOD 11-2 check digit
verification.
}
\references{
\url{https://support.orcid.org/hc/en-us/articles/360006897674-Structure-of-the-ORCID-Identifier}
}
