% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{is_sedol}
\alias{is_sedol}
\title{Check if an argument is a valid SEDOL string}
\usage{
is_sedol(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid SEDOL string, \code{FALSE} otherwise.
}
\description{
Validates SEDOL format including weighted check digit verification.
}
\examples{
is_sedol("0263494")
}
\references{
\url{https://en.wikipedia.org/wiki/SEDOL}
}
