% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_ulid}
\alias{is_ulid}
\title{Check if an argument is a ULID string}
\usage{
is_ulid(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid ULID string, \code{FALSE} otherwise.
}
\description{
Check if an argument is a ULID string
}
\examples{
is_ulid("01ARZ3NDEKTSV4RRFFQ69G5FAV")
}
\references{
\url{https://github.com/ulid/spec}
}
