% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_url}
\alias{is_url}
\title{Check if an argument is url string}
\usage{
is_url(x)
}
\arguments{
\item{x}{(`any`)\cr
Object to check.}
}
\value{
`TRUE` if `x` is a valid URL string, `FALSE` otherwise.
}
\description{
Check if an argument is url string
}
\examples{
is_url("https://example.com")
}
