% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{is_uuid}
\alias{is_uuid}
\title{Check if an argument is an UUID string}
\usage{
is_uuid(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}
}
\value{
\code{TRUE} if \code{x} is a valid UUID string, \code{FALSE} otherwise.
}
\description{
Check if an argument is an UUID string
}
\examples{
is_uuid("550e8400-e29b-41d4-a716-446655440000")
}
